Distributions: move non-central into continuous
